requirements.txt
# Framework Principal
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Base de Datos
psycopg2-binary>=2.9.9

# Autenticación y Seguridad
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1

# Utilidades
python-decouple>=3.8
django-filter>=23.5
Pillow>=10.1.0
